+++ vove-aplication/src/main/java/sptech/school/voveaplication/domain/arquivo/repository/ArquivoRepository.java
@@ -3,5 +3,5 @@ package sptech.school.voveaplication.domain.arquivo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sptech.school.voveaplication.domain.arquivo.Arquivo;
 
-public interface ArquivoRepository extends JpaRepository<Arquivo, Integer> {
+public interface ArquivoRepository extends JpaRepository<Arquivo, Long> {
 }
